Create Cubaka Thesis
This is a Code written in GAP for my master's thesis. I am a current student at African Institite for Mathematical Sciences (AIMS Ghana) and working on strategies of word shortening. I provide codes to shorten word using built-in function in GAP and test them on the Rubik's Cube group for words of length 40, 100, and 200.